refactor(transport)!: split transport.py into a focused package (#163)
* refactor(transport): split transport.py into a focused package

Move the four logical concerns of the old 1.3kloc transport.py into
their own modules:

  goggles/_core/transport/
      __init__.py    # re-exports the flat surface tests/routing rely on
      _frames.py     # wire format, env knobs, shm helpers
      _endpoints.py  # _Endpoint protocol + Unix/Windows + selector
      _protocol.py   # Transport Protocol
      _local.py      # LocalTransport + _SendItem + _SENTINEL

`from goggles._core.transport import X` keeps working for every
public and underscore-prefixed symbol the test suite touched
(LocalTransport, Transport, _pack_small_frame, _unpack_small,
_HEADER_FMT, _IS_WINDOWS, _MSG_SMALL, _SHM_NAME_PREFIX,
_default_socket_path, _next_shm_name, _reap_orphan_shm,
_try_unlink_shm, _user_tag, ...). No behaviour change; git tracks
transport.py → _local.py as a rename so blame still resolves on
the LocalTransport body.

Closes #160.


* docs(transport): correct claim about why Transport is runtime_checkable

The original docstring said `runtime_checkable` was needed for the
routing singleton's rebuild check, but routing.py gates on the
`is_running` attribute directly — no `isinstance` involved. The
decorator is still useful for user-side `isinstance(x, Transport)`
checks; just say so.


* docs(transport): move package layout from __init__ docstring to transport.md

albonomi02 asked for the package architecture explanation to live in a
docs file rather than as a long module docstring. Add docs/guides/transport.md
covering the public surface, internal layout, and flat re-export
rationale; slim __init__.py's docstring to a one-liner that points there.

Update architecture.md so the package-layout tree, section-1 routing
header, section-4 transport section, and the "Where to make changes"
table all reflect the transport.py to transport/ split. Add a
guides-table row in docs/index.md so the new file is discoverable
from the hub.


* refactor(transport): drop unneeded type: ignore on _endpoint() returns

The structural Protocol match is recognized by basedpyright, so the
return-value ignores were noise. Removed both.


---------
